moving functions into more appropriate locations
