+++ src/wf_convert/mod.rs
@@ -1,6 +1,5 @@
 use crate::wf_core::basic_message::BasicMessage;
-use crate::wf_core::segment::MessageSegment;
-use crate::wf_field::get_message_body;
+use crate::wf_core::segment::{MessageSegment, get_message_body};
 
 pub trait FieldValue: AsRef<str> + Into<String> {}
 impl<T> FieldValue for T where T: AsRef<str> + Into<String> {}
